pyproject.toml
[project]
name = "nodriverheadless"
version = "0.1.0"
description = "Ghost-Browser automation suite"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "zendriver",
    "stealth-browser-controller",
    "numpy",
    "scipy",
    "requests",
    "python-dotenv",
    "opencv-python-headless",
    "pyyaml",
    "aiohttp[socks]>=3.13.3",
    "aiohttp-socks>=0.11.0",
    "pysocks>=1.7.1",
]

[tool.uv]
managed = true
package = false
